Update 608037-PHIAnnotationDeIdentification.md
> and resolves potential annotation conflicts

the Deidentifier doesn't really do this, yet.

Also, I think "token" should be better defined (or at least it should be acknowledged that we have some specific way of defining tokens). Tokenization is not trivial!

I suggested the change of "Driver" -> "collaborator" that I did in my other pulls. We should only make this change if it is consistent across the wiki.