Cascadianblue patch 2 (#69)
* Update 607935-Tasks.md

I removed "reproducible" because I'm not sure it's an appropriate term for NLP tools in this context. I'm open to being corrected if there's a good explanation of what it's supposed to mean.

I'm not sure what:

> (see Section OpenAPI TODO)

means.

A few key terms referring to groups/individuals are capitalized (Driver, Developer, Data Hosting Sites). This seems odd and I'm getting rid of the capitalization unless there's a precedent for it. I also tried changing the word "Driver" to "collaborator" and "Developer" to "participant".

* Review feedback

Co-authored-by: Connor Boyle <connor.boyle@sagebase.org>